chore: ARQUI_TPd

+++ arquitectura_computadoras/tps/tp05-pretpe/barebones/Kernel/include/sysCalls.h
@@ -1,6 +1,6 @@
 #ifndef SYSCALLS_H
 #define SYSCALLS_H
 
-size_t write(int fd, char * buff, int length);
+int write(int fd, char * buff, int length);
 
 #endif
